feat(naivi-dom): Rust frame core — decode/apply/reset/rejection (U6)
The Rust side of plan 074 stage 2 lands: frames arrive as one binary blob,
are decoded and applied as a whole transaction, and rejections are queued
for the host to surface via frame_rejected (KTD1/KTD3/KD8, AE5).

- frame.rs (new): FrameOp (virtual-id addressed) + FrameDecoder (single-pass
  borrowing decoder over the @naivi/protocol wire format) + FrameApplier
  (order-aware validation: id live iff in the persistent virtual map or
  created earlier in the frame; whole-frame rejection, no partial DOM)
- ops.rs: OpsCore gains the shared virtual-id map (KTD7), virtual-aware
  methods (create_*_v / set_*_v / append/insert/replace/remove/bind/unbind_v
  / attach_root_v), bind_event_v writes the VIRTUAL id into data-naivi-id
  (KTD2), and reset() drops the tree + map + bindings (R15)
- document.rs: NaiviDocument shares the virtual map, owns the
  frame_rejected(seq, reason) queue, and gains flush_frame(bytes) + reset()
- events.rs: NaiviEvent.node is now the u32 virtual id; the handler
  reverse-looks-up data-naivi-id at queue time (KTD2)
- ffi.rs + wasm/native sinks: QueuedEvent/NaiviEvent delivery now carries
  the virtual id (u32)
- tests: frame.rs (7 tests: legal frame, transactional rejection, duplicate
  create, remove-invalidate, reset, malformed, event virtual-id delivery);
  ops.rs event tests migrated to the virtual-id path

Verification: cargo test -p naivi-dom (28 green); cargo check --workspace.

+++ packages/naivi-dom/src/ffi.rs
@@ -24,8 +24,9 @@ use std::str::FromStr;
 /// A single event queued for the guest callback.
 #[derive(Clone, Debug)]
 pub struct QueuedEvent {
-    /// The bound node's id (as `u64`; crosses to JS as a `f64` number).
-    pub node: u64,
+    /// The bound node's **virtual id** (JS-assigned u32; crosses to JS as a
+    /// `f64` number). Replaces the old u64 blitz `NodeId` (U6/KTD2).
+    pub node: u32,
     /// The protocol `u8` event kind.
     pub kind: u8,
     /// Client (viewport-relative) coordinates, when present.
